Update RTL tests
