+++ testing/component-testing-rtl/src/__tests__/2-user-interaction.test.js
